SubArray: Fix incorrect assertions in 2-level MLC code.
If all 512 bits in a cache line are changed, an assertion files comparing
the number of changed bits to the size of the request data. If there are
a combination of only 00->11 or 11->00 bit changes, another assertion is
triggered. This is changed set the write time to the maximum pulse length
of 11 or 00 writes.